pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepspot2cell"
version = "0.1.0"
description = "DeepSpot2Cell: virtual single-cell spatial transcriptomics from H&E images"
authors = [{ name = "DeepSpot2Cell Contributors" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "lightning>=2.2.0",
    "torchmetrics>=1.3.0",
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.2.0",
    "numba>=0.56.0",
    "llvmlite>=0.39.0",
    "anndata>=0.8.0",
    "scanpy>=1.9.0",
    "squidpy>=1.2.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.65.0",
    "h5py>=3.8.0",
    "pyyaml>=6.0",
    "shapely>=2.0",
    "pyvips>=2.2.1",
    "opencv-python>=4.9.0",
    "pillow>=10.0.0",
    "huggingface_hub>=0.23.0",
    "transformers>=4.42.0",
    "timm>=0.9.12",
    "plotnine>=0.12.0",
    "einops>=0.7.0",
    "hest>=0.2.0",
]

[project.optional-dependencies]
dev = ["black", "isort", "nbstripout"]

[tool.setuptools.packages.find]
include = ["deepspot2cell", "deepspot2cell.*"]
